+++ nested_cv_shuf_new_name.py
@@ -12,7 +12,7 @@ import time
 import subprocess
 import sys
 
-print("WARNING THIS IS AN EDITED SCRIPT - Ciaran Kelly 2021 \n Edited with permission under liscence")
+print("WARNING THIS IS AN EDITED SCRIPT - Ciaran Kelly 2021 \n Edited with permission under liscence \n Shuf version")
 set_size = 506    
 
 def load_data(data):
@@ -24,20 +24,26 @@ def load_data(data):
         #x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.8, random_state=42)
         return x, y #x_train, y_train, x_test, y_test
 
-def bash_script(train_index, test_index, train_names, test_names):
-        with open("name_vector_train.txt", 'w') as f:
-            for item in train_names:
-                f.write("%s %s\n" % (item, item))
-        with open("name_vector_test.txt", 'w') as f:
-            for item in test_names:
-                f.write("%s %s\n" % (item, item))        
-        subprocess.run(["/external_storage/ciaran/machine_learning2/bash_script_shuf.sh"], shell=True) 
-        time.sleep(60)
-        while not os.path.exists('train_raw_plink_shuf.raw'):
-                print("Waiting")
-                time.sleep(20)
-        new_X_train , new_y_train = load_data('train_raw_plink_shuf.raw') #made from bash_script.sh
-        new_X_test , new_y_test  = load_data('test_raw_plink_shuf.raw')
+def bash_script(train_index, test_index, train_names, test_names, outer_count, inner_count, outer=False):
+        if outer==True:
+            foo='out'
+        else:
+            foo='in'
+        if not os.path.exists('train_raw_plink_shuf_' + str(outer_count) + '_in_' + str(inner_count) + '_' + foo + '.raw'):
+            print("SETTING OFF CUSTOM BASH SCRIPT")
+            with open("name_vector_train.txt", 'w') as f:
+                for item in train_names:
+                    f.write("%s %s\n" % (item, item))
+            with open("name_vector_test.txt", 'w') as f:
+                for item in test_names:
+                    f.write("%s %s\n" % (item, item))        
+
+            subprocess.run(["/external_storage/ciaran/machine_learning2/bash_script_shuf.sh", str(outer_count), str(inner_count), foo]) 
+        #while not os.path.exists('train_raw_plink_shuf_' + str(outer_count) + '_in_' + str(inner_count) + '.raw'):
+        while not os.path.exists('test_raw_plink_shuf_' + str(outer_count) + '_in_' + str(inner_count) + '_' + foo + '.raw'):
+            time.sleep(20)
+        new_X_train , new_y_train = load_data('train_raw_plink_shuf_' + str(outer_count) + '_in_' + str(inner_count) + '_' + foo + '.raw') #made from bash_script.sh
+        new_X_test , new_y_test  = load_data('test_raw_plink_shuf_' + str(outer_count) + '_in_' + str(inner_count) + '_' + foo + '.raw')
         return new_X_train, new_X_test, new_y_train, new_y_test
                               
 
@@ -98,7 +104,8 @@ class NestedCV():
             average other than 'binary'
     '''
 
-    def __init__(self, name_list, model, params_grid, outer_kfolds, inner_kfolds, n_jobs = 1, cv_options={}):
+    def __init__(self, model_name, name_list, model, params_grid, outer_kfolds, inner_kfolds, n_jobs = 1, cv_options={}):
+        self.model_name = model_name
         self.name_list = name_list
         self.model = model
         self.params_grid = params_grid
@@ -262,12 +269,15 @@ class NestedCV():
         
         outer_cv = KFold(n_splits=self.outer_kfolds, shuffle=True, random_state=42)
         inner_cv = KFold(n_splits=self.inner_kfolds, shuffle=True, random_state=42)
-        inner_count = 0 ; outer_count = 0
+        outer_count = 1
         outer_scores = []
         variance = []
         best_inner_params_list = []  # Change both to by one thing out of key-value pair
         best_inner_score_list = []
-
+        print("ADBUCE")
+        print(self.model_name)
+        #print(type(self.model)._keras_api_names)
+        #modeltype = str(type(self.model).__name__)
         # Split X and y into K-partitions to Outer CV
         for (i, (train_index, test_index)) in enumerate(outer_cv.split(X, y)):
             log.debug(
@@ -278,7 +288,8 @@ class NestedCV():
             best_inner_params = {}
             best_inner_score = None
             search_scores = []
-
+            inner_count = 0
+            outer=False
             # Split X_train_outer and y_train_outer into K-partitions to be inner CV
             for (j, (train_index_inner, test_index_inner)) in enumerate(inner_cv.split(X_train_outer, y_train_outer)):
                 log.debug(
@@ -287,7 +298,7 @@ class NestedCV():
                 #y_train_inner, y_test_inner = y_train_outer[train_index_inner], y_train_outer[test_index_inner]
                 inner_count += 1 ; print("INNER COUNT NO. ", str(inner_count))
                 inner_train_names, inner_test_names = outer_train_names[train_index_inner], outer_train_names[test_index_inner]
-                X_train_inner, X_test_inner, y_train_inner, y_test_inner = bash_script(train_index_inner, test_index_inner, inner_train_names, inner_test_names)
+                X_train_inner, X_test_inner, y_train_inner, y_test_inner = bash_script(train_index_inner, test_index_inner, inner_train_names, inner_test_names,outer_count, inner_count)
                 if self.recursive_feature_elimination:
                         X_train_inner, X_test_inner = self._fit_recursive_feature_elimination(
                                     X_train_inner, y_train_inner, X_test_inner)
@@ -303,12 +314,14 @@ class NestedCV():
             
             best_inner_params_list.append(best_inner_params)
             best_inner_score_list.append(best_inner_score)
-            outer_count += 1 ; print("OUTER COUNT NO. ", str(outer_count))
+            #inner_count = 0
+            print("OUTER COUNT NO. ", str(outer_count))
             # Fit the best hyperparameters from one of the K inner loops
             self.model.set_params(**best_inner_params)
-            X_train_outer, X_test_outer, y_train_outer, y_test_outer = bash_script(train_index, test_index, outer_train_names, inner_test_names)
+            X_train_outer, X_test_outer, y_train_outer, y_test_outer = bash_script(train_index, test_index, outer_train_names, inner_test_names, outer_count, inner_count, outer=True)
+            outer = False
             self.model.fit(X_train_outer, y_train_outer.ravel())
-            
+            outer_count += 1 
             # Get score and prediction
             score,pred = self._predict_and_score(X_test_outer, y_test_outer.ravel())
             outer_scores.append(self._transform_score_format(score))
